requirement.txt
langchain
langchain_community
langchain_core==0.2.0
langchain_openai
python-dotenv
streamlit
beautifulsoup4
